Moved RDS in modules dir
